Deprecated ddtrace.http, merged ddtrace.http.headers into trace_utils
