+++ ddtrace/http/__init__.py
@@ -1,3 +1,5 @@
+from ddtrace.utils.deprecation import deprecation
+
 from .headers import store_request_headers
 from .headers import store_response_headers
 
@@ -6,3 +8,9 @@ __all__ = [
     "store_request_headers",
     "store_response_headers",
 ]
+
+deprecation(
+    name="ddtrace.http",
+    message="The http module has been merged into ddtrace.contrib.trace_utils",
+    version="1.0.0",
+)
